단축URL 생성 서비스: 예외 핸들러 추가, 테스트 케이스 추가, 리팩토링

+++ shortenurlservice/src/main/java/kr/co/shortenurlservice/presentation/ShortenUrlDTO.java
@@ -4,14 +4,14 @@ import kr.co.shortenurlservice.domain.ShortenUrl;
 public class ShortenUrlDTO {
     private String originalUrl;
     private String shortenUrlKey;
-    private int redirectCount;
+    private Long redirectCount;
 
     public ShortenUrlDTO() { }
     public ShortenUrlDTO(String originalUrl) {
         this.originalUrl = originalUrl;
     }
 
-    public ShortenUrlDTO(String originalUrl, String shortenUrlKey, int redirectCount) {
+    public ShortenUrlDTO(String originalUrl, String shortenUrlKey, Long redirectCount) {
         this.originalUrl = originalUrl;
         this.shortenUrlKey = shortenUrlKey;
         this.redirectCount = redirectCount;
@@ -25,7 +25,7 @@ public class ShortenUrlDTO {
         return shortenUrlKey;
     }
 
-    public int getRedirectCount() { return redirectCount; };
+    public Long getRedirectCount() { return redirectCount; };
 
     public static ShortenUrlDTO toDTO(ShortenUrl shortenUrl) {
         return new ShortenUrlDTO(
